OE-8015 - Remove tabindex html attribute from eyedraw canvas
The tabindex attribute was interfering with the tabbing order of other
elements in Examination edit/create mode. Changing the value of tabindex
from "1" to "0" also was a potential fix for the problem, and still
allowed the canvas to be focused on. However, since there seems to be no
reason for the user to focus the keyboard on the eyedraw canvas, the
chosen solution in this commit is to entirely remove the tabindex
attribute.