OE-8733 Added unique group classes to lenses in eyedraw.
Specified a unique group name, which indicates that any time a doodle
with this unique group is present in an drawing, it wants to be the only
doodle of that type in the drawing. By itself this change does not
enforce this property, it is enforced in OpenEyes branch fix/OE-8733.